short => int, fix for bugzilla 4130?

+++ src/backend/cc.h
@@ -1376,7 +1376,7 @@ struct Symbol
                                 // SCauto,SCparameter,SCtmp,SCregpar,SCregister
     // CODGEN
 #if TX86
-    short Sseg;                 // segment index
+    int Sseg;                   // segment index
 #endif
     int Sweight;                // usage count, the higher the number,
                                 // the more worthwhile it is to put in
